requirements.txt
# Note: LangChain frequently changes their API. Modify the versions & dependencies accordingly.
langchain>=0.2
langchain-community>=0.2
langchain-chroma>=0.1
langchain-text-splitters>=0.2
langchain-groq>=0.1
langchain-huggingface>=0.1
chromadb>=0.5
unstructured>=0.15
beautifulsoup4>=4.12
lxml>=5.0
python-dotenv>=1.0
streamlit>=1.36
validators>=0.28
pytest>=8.0
